refactor(blorktools): rig tooltip manager
